Made the drag zone not clickable, fixed a bug and added comments.
Refactored drag and drop and file input handling logic so that the drag zone cannot be clicked to upload a file. Fixed a bug where users can drag a file in while another file is being uploaded. Added detailed JSDoc comments for functions and smaller comments in code to improve code documentation.